remove Transformer notebook
