+++ aeon/transformations/_legacy/summarize.py
@@ -5,19 +5,12 @@ __all__ = ["SummaryTransformer", "WindowSummarizer"]
 
 import numpy as np
 import pandas as pd
-from deprecated.sphinx import deprecated
 from joblib import Parallel, delayed
 
 from aeon.transformations.base import BaseTransformer
 from aeon.utils.multiindex import flatten_multiindex
 
 
-# TODO: remove in v0.11.0
-@deprecated(
-    version="0.10.0",
-    reason="WindowSummarizer will be removed in version 0.11.0.",
-    category=FutureWarning,
-)
 class WindowSummarizer(BaseTransformer):
     """
     Transformer for extracting time series features.
@@ -132,55 +125,6 @@ class WindowSummarizer(BaseTransformer):
         Contains all transformed columns as well as non-transformed columns.
         The raw inputs to transformed columns will be dropped.
     self: reference to self
-
-    Examples
-    --------
-    >>> import pandas as pd
-    >>> from aeon.transformations.summarize import WindowSummarizer
-    >>> from aeon.datasets import load_airline, load_longley
-    >>> from aeon.forecasting.naive import NaiveForecaster
-    >>> from aeon.forecasting.base import ForecastingHorizon
-    >>> from aeon.forecasting.compose import ForecastingPipeline
-    >>> from aeon.forecasting.model_selection import temporal_train_test_split
-    >>> y = load_airline()
-    >>> kwargs = {
-    ...     "lag_feature": {
-    ...         "lag": [1],
-    ...         "mean": [[1, 3], [3, 6]],
-    ...         "std": [[1, 4]],
-    ...     }
-    ... }
-    >>> transformer = WindowSummarizer(**kwargs)
-    >>> y_transformed = transformer.fit_transform(y)
-
-        Example with transforming multiple columns of exogeneous features
-    >>> y, X = load_longley()
-    >>> y_train, y_test, X_train, X_test = temporal_train_test_split(y, X)
-    >>> fh = ForecastingHorizon(X_test.index, is_relative=False)
-    >>> # Example transforming only X
-    >>> pipe = ForecastingPipeline(
-    ...     steps=[
-    ...         ("a", WindowSummarizer(n_jobs=1, target_cols=["POP", "GNPDEFL"])),
-    ...         ("b", WindowSummarizer(n_jobs=1, target_cols=["GNP"], **kwargs)),
-    ...         ("forecaster", NaiveForecaster(strategy="drift")),
-    ...     ]
-    ... )
-    >>> pipe_return = pipe.fit(y_train, X_train)
-    >>> y_pred1 = pipe_return.predict(fh=fh, X=X_test)
-
-        Example with transforming multiple columns of exogeneous features
-        as well as the y column
-    >>> Z_train = pd.concat([X_train, y_train], axis=1)
-    >>> Z_test = pd.concat([X_test, y_test], axis=1)
-    >>> pipe = ForecastingPipeline(
-    ...     steps=[
-    ...         ("a", WindowSummarizer(n_jobs=1, target_cols=["POP", "TOTEMP"])),
-    ...         ("b", WindowSummarizer(**kwargs, n_jobs=1, target_cols=["GNP"])),
-    ...         ("forecaster", NaiveForecaster(strategy="drift")),
-    ...     ]
-    ... )
-    >>> pipe_return = pipe.fit(y_train, Z_train)
-    >>> y_pred2 = pipe_return.predict(fh=fh, X=Z_test)
     """
 
     _tags = {
@@ -594,12 +538,6 @@ def _check_quantiles(quantiles):
     return quantiles
 
 
-# TODO: remove in v0.11.0
-@deprecated(
-    version="0.10.0",
-    reason="SummaryTransformer will be removed in version 0.11.0.",
-    category=FutureWarning,
-)
 class SummaryTransformer(BaseTransformer):
     """Calculate summary value of a time series.
 
@@ -633,14 +571,6 @@ class SummaryTransformer(BaseTransformer):
     -----
     This provides a wrapper around pandas DataFrame and Series agg and
     quantile methods.
-
-    Examples
-    --------
-    >>> from aeon.transformations.summarize import SummaryTransformer
-    >>> from aeon.datasets import load_airline
-    >>> y = load_airline()
-    >>> transformer = SummaryTransformer()
-    >>> y_mean = transformer.fit_transform(y)
     """
 
     _tags = {
@@ -742,12 +672,6 @@ class SummaryTransformer(BaseTransformer):
         return [params1, params2, params3]
 
 
-# TODO: remove in v0.11.0
-@deprecated(
-    version="0.10.0",
-    reason="PlateauFinder will be removed in version 0.11.0.",
-    category=FutureWarning,
-)
 class PlateauFinder(BaseTransformer):
     """
     Plateau finder transformer.
@@ -833,12 +757,6 @@ class PlateauFinder(BaseTransformer):
         return Xt
 
 
-# TODO: remove in v0.11.0
-@deprecated(
-    version="0.10.0",
-    reason="FittedParamExtractor will be removed in version 0.11.0.",
-    category=FutureWarning,
-)
 class FittedParamExtractor(BaseTransformer):
     """Fitted parameter extractor.
