+++ aeon/transformations/tests/test_plateaufinder.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from aeon.transformations.summarize import PlateauFinder
+from aeon.transformations._legacy.summarize import PlateauFinder
 
 
 @pytest.mark.parametrize("value", [np.nan, -10, 10, -0.5, 0.5])
